requirements.txt
# AI & Image Processing
ultralytics>=8.0.0
opencv-python
numpy

# GUI
PyQt6

# Database & Config
mysql-connector-python
pyyaml

# System
requests